+++ mrpt_tps_astar_planner/package.xml
@@ -1,56 +1,38 @@
 <?xml version="1.0"?>
-<package format="2">
-  <name>tps_astar_nav</name>
-  <version>0.0.1</version>
+<package format="3">
+  <name>mrpt_tps_astar_planner</name>
+  <version>1.0.3</version>
   <description>ROS Navigator with A* in TP-Space Engine</description>
-  <maintainer email="shravansomashekara.rai@gmail.com">Shravan_S_Rai</maintainer>
-  <license>BSD-3-Clause license</license>
-  <url type="website">TODO</url>
-  <author email="shravansomashekara.rai@gmail.com">Shravan_S_Rai</author>
 
-  <!--Dependencie  -->
-  <doc_depend>doxygen</doc_depend>
-  <buildtool_depend>catkin</buildtool_depend>
+  <maintainer email="jlblanco@ual.es">Jose-Luis Blanco-Claraco</maintainer>
+  <maintainer email="shravansomashekara.rai@gmail.com">Shravan S Rai</maintainer>
+
+  <author email="shravansomashekara.rai@gmail.com">Shravan_S_Rai</author>
+  <author email="jlblanco@ual.es">Jose-Luis Blanco-Claraco</author>
+
+  <license>BSD</license>
+
+  <url type="website">https://github.com/mrpt-ros-pkg/mrpt_navigation</url>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
   <depend>mrpt2</depend>
-  <depend>actionlib</depend>
-  <depend>actionlib_msgs</depend>
-  <depend>dynamic_reconfigure</depend>
+
+  <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
+  <depend>nav_msgs</depend>
+  <depend>sensor_msgs</depend>
+  <depend>geometry_msgs</depend>
+  <depend>mrpt_msgs</depend>
   <depend>tf2</depend>
   <depend>tf2_ros</depend>
   <depend>tf2_geometry_msgs</depend>
   <depend>visualization_msgs</depend>
-  <build_depend>mrpt_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>rospy</build_depend>
-  <build_depend>sensor_msgs</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>nav_msgs</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>message_generation</build_depend>
 
-  <build_export_depend>roscpp</build_export_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <build_export_depend>sensor_msgs</build_export_depend>
-  <build_export_depend>nav_msgs</build_export_depend>
-  <build_export_depend>geometry_msgs</build_export_depend>
-  <build_export_depend>std_msgs</build_export_depend>
-  <build_export_depend>mrpt_msgs</build_export_depend>
-  <build_export_depend>tf</build_export_depend>
-  <build_export_depend>message_generation</build_export_depend>
-  
-  <exec_depend>roscpp</exec_depend>
-  <exec_depend>rospy</exec_depend>
-  <exec_depend>sensor_msgs</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>tf</exec_depend>
-  <exec_depend>message_runtime</exec_depend>
-  <exec_depend>nav_msgs</exec_depend>
-  <exec_depend>geometry_msgs</exec_depend>
-  <exec_depend>mrpt_msgs</exec_depend>
+  <depend>ament_lint_common</depend>
+  <depend>ament_lint_auto</depend>
 
-  <!-- The export tag contains other, unspecified, tags -->
   <export>
-    <!-- Other tools can request additional information be placed here -->
-
+    <build_type>ament_cmake</build_type>
   </export>
 </package>
